+++ rtw-serving/src/main/kotlin/com/javacafe/rtwserving/RtwServingApplication.kt
@@ -1,0 +1,11 @@
+package com.javacafe.rtwserving
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class RtwServingApplication
+
+fun main(args: Array<String>) {
+    runApplication<RtwServingApplication>(*args)
+}
