+++ rtw-serving/src/main/kotlin/com/javacafe/rtwserving/common/dto/ApiResponse.kt
@@ -1,0 +1,30 @@
+package com.javacafe.rtwserving.common.dto
+
+import java.time.LocalDateTime
+
+data class ApiResponse<T>(
+    val data: T? = null,
+    val message: String,
+    val timestamp: LocalDateTime = LocalDateTime.now()
+) {
+    companion object {
+        fun <T> success(
+                data: T,
+                message: String = "성공"
+        ): ApiResponse<T> {
+            return ApiResponse(
+                data = data,
+                message = message
+            )
+        }
+
+        fun error(
+            message: String = "오류가 발생했습니다"
+        ): ApiResponse<Nothing> {
+            return ApiResponse(
+                data = null,
+                message = message
+            )
+        }
+    }
+}
