+++ rtw-serving/src/main/kotlin/com/javacafe/rtwserving/common/exception/GlobalException.kt
@@ -1,0 +1,7 @@
+package com.javacafe.rtwserving.common.exception
+
+open class GlobalException (
+    val exceptionCode: ExceptionCode,
+    override val message: String? = exceptionCode.message,
+    val extra: Map<String, Any>? = null,
+) : RuntimeException(message ?: exceptionCode.message)
