+++ rtw-serving/src/main/kotlin/com/javacafe/rtwserving/config/ElasticsearchConfig.kt
@@ -1,0 +1,10 @@
+package com.javacafe.rtwserving.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories
+
+@Configuration
+@EnableElasticsearchRepositories(
+    basePackages = ["com.javacafe.rtwserving.domain.*.repository"]
+)
+class ElasticsearchConfig
