+++ rtw-serving/src/main/resources/application.yml
@@ -1,0 +1,18 @@
+server:
+  port: 8080
+
+spring:
+  application:
+    name: rtw-serving
+    
+  # Elasticsearch 설정
+  elasticsearch:
+    uris: http://localhost:9200
+    connection-timeout: 10s
+    socket-timeout: 60s
+
+# 로깅 설정 (개발용)
+logging:
+  level:
+    com.javacafe: DEBUG
+    org.springframework.data.elasticsearch: INFO
