+++ settings.gradle.kts
@@ -1,1 +1,4 @@
 rootProject.name = "realtime-winnmin"
+
+include("rtw-collector")
+include("rtw-serving")
