--- src/main/kotlin/com/javacafe/realtimewinnmin/RealtimeWinnminApplication.kt
@@ -1,11 +1,0 @@
-package com.javacafe.realtimewinnmin
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class RealtimeWinnminApplication
-
-fun main(args: Array<String>) {
-    runApplication<RealtimeWinnminApplication>(*args)
-}
