--- src/main/kotlin/com/javacafe/realtimewinnmin/common/config/AopConfig.kt
@@ -1,8 +1,0 @@
-package com.javacafe.realtimewinnmin.common.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.EnableAspectJAutoProxy
-
-@Configuration
-@EnableAspectJAutoProxy
-class AopConfig
