--- src/main/kotlin/com/javacafe/realtimewinnmin/infrastructure/config/ElasticsearchConfig.kt
@@ -1,10 +1,0 @@
-package com.javacafe.realtimewinnmin.infrastructure.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories
-
-@Configuration
-@EnableElasticsearchRepositories(
-    basePackages = ["com.javacafe.realtimewinnmin.domain.*.repository"]
-)
-class ElasticsearchConfig
