--- src/main/resources/application.yml
@@ -1,18 +1,0 @@
-spring:
-  application:
-    name: realtime-winnmin
-    
-  # Elasticsearch 설정
-  elasticsearch:
-    uris: http://localhost:9200
-    connection-timeout: 10s
-    socket-timeout: 60s
-    
-server:
-  port: 8080
-
-# 로깅 설정 (개발용)
-logging:
-  level:
-    com.javacafe: DEBUG
-    org.springframework.data.elasticsearch: INFO
