--- src/test/kotlin/com/javacafe/realtimewinnmin/RealtimeWinnminApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.javacafe.realtimewinnmin
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class RealtimeWinnminApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
